pyproject.toml
[project]
name = "lounge"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "moi_moisture", email = "138105872+momok47@users.noreply.github.com" }
]
dependencies = [
    "numpy>=2.0.0",
    "matplotlib>=3.8.0",
    "japanize-matplotlib>=1.1.3",
    "sox>=1.5.0",
    "pyyaml>=6.0.2",
    "ipykernel>=6.0.0",
    "mecab-python3>=1.0.10",
    "unidic>=1.1.0",
    "google-cloud>=0.34.0",
    "django>=4.2,<5.0",
    "google-cloud-storage>=3.1.0",
    "google-cloud-speech>=2.32.0",
    "google>=3.0.0",
    "pandas>=2.2.0",
    "torch>=2.8.0",
    "transformers>=4.55.2",
    "fugashi>=1.4.0",
    "unidic-lite>=1.0.8",
    "mecab>=0.996.11",
    "ipadic>=1.0.0",
    "openai>=1.101.0",
    "python-dotenv>=1.0.0",
    "pyaudio>=0.2.14",
    "dimod>=0.12.0",
    "dwave-neal>=0.6.0",
    "dwave-ocean-sdk>=8.0.0",
    "datasets>=4.0.0",
    "scikit-learn>=1.7.1",
    "tiktoken>=0.11.0",
    "blobfile>=3.0.0",
    "tqdm>=4.67.1",
    "wandb>=0.21.0",
    "sentencepiece>=0.2.0",
    "peft>=0.17.0",
    "trl>=0.21.0",
    "bitsandbytes>=0.42.0",
    "accelerate>=1.10.0",
    "psutil>=6.0.0",
]
readme = "README.md"
requires-python = ">= 3.13"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/lounge"]
